Major Restructure; Classes fixes;
