+++ src/manegments_system/__old/opti_with_.py
