+++ src/manegments_system/__old/optimization with class.py
