+++ src/manegments_system/menu.py
@@ -1,0 +1,10 @@
+def show_menu():
+    print("\nWould you like to --")
+    print("1. See current Users")
+    print("2. Add User")
+    print("3. Delete User")
+    print("4. Show a User")
+    print("5. Save Database")
+    print("6. More Options")
+    print("7. Quit")
+
